requirements.txt

aiosignal==1.3.1
asgiref==3.8.1
async-lru==2.0.4
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.6
frozenlist==1.4.1
h11==0.14.0
idna==3.7
image==1.5.33
logging==0.4.9.6
lxml==5.2.2
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.1
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
termcolor==2.4.0
trio==0.25.1
trio-websocket==0.11.1
tweepy==4.14.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
